Fixes and improvements to example/nce-loss - No functional change
* pylint conformity
* Executing with python 2.7 and 3.6 fixed
* Replaced Model.FeedForward with Module interface to get fix deprecation warning
* README.md improved to include better information on the dataset as well as example scripts
* Pulled code for reading data and creating network into their own file for better organization and to avoid variable name clashes
